+++ resolwe/flow/executors/zeromq_utils.py
@@ -1,14 +1,11 @@
 """Utils for working with zeromq."""
 
 import json
-import os
 from logging import Logger
-from threading import Lock
 from typing import Any, Optional, Tuple
 
 import zmq
 import zmq.asyncio
-from zmq.auth.asyncio import AsyncioAuthenticator
 
 from .socket_utils import BaseCommunicator, PeerIdentity
 
@@ -78,26 +75,3 @@ class ZMQCommunicator(BaseCommunicator):
             async_zmq_send_data,
             async_zmq_receive_data,
         )
-
-
-class ZMQAuthenticator(AsyncioAuthenticator):
-    """The singleton authenticator."""
-
-    _instance = None
-    _instance_lock = Lock()
-    _instance_pid: int | None = None
-
-    @classmethod
-    def has_instance(cls):
-        """Check if the instance exists."""
-        return not (cls._instance is None or cls._instance_pid != os.getpid())
-
-    @classmethod
-    def instance(cls, context=None):
-        """Return a global ZMQAuthenticator instance."""
-        if not cls.has_instance():
-            with cls._instance_lock:
-                if not cls.has_instance():
-                    cls._instance = cls(context=context)
-                    cls._instance_pid = os.getpid()
-        return cls._instance